add validations to items app

+++ api/apps/equipment_app/controllers/equipment_controller.py
@@ -2,6 +2,15 @@ from flask import g, current_app
 
 
 class EquipmentController:
+    @classmethod
+    def _user_has_privileges(cls, trip_id=None, role_id=None, owner_id=None):
+        user = cls._get_user(g.user_id)
+        trip = cls._get_trip(trip_id)
+        user_has_role = role_id in (role.id for role in user.roles)
+        is_item_owner = owner_id == user.user_id
+        is_admin = user == trip.admin
+        return user_has_role or is_item_owner or is_admin
+
     @staticmethod
     def _get_eq(equipment_id):
         eq = current_app.models.Equipment.get_equipment_by_id(equipment_id)
@@ -27,23 +36,23 @@ class EquipmentController:
 
     @classmethod
     def update_equipment(cls, equipment_id, data):
-        new_equipment = current_app.models.Equipment.update_equipment(equipment_id, data)
-        return new_equipment
+        item = cls._get_eq(equipment_id)
+        if cls._user_has_privileges(item.trip_id, item.role_id, item.owner_id):
+            response = current_app.models.Equipment.update_equipment(equipment_id, data)
+            return response, 201
+        return 'You dont have rights', 402
 
     @classmethod
     def delete_equipment(cls, equipment_id):
-        equipment_data = current_app.models.Equipment.delete_equipment(equipment_id)
-        return equipment_data
+        item = cls._get_eq(equipment_id)
+        if cls._user_has_privileges(item.trip_id, item.role_id, item.owner_id):
+            response = current_app.models.Equipment.delete_equipment(equipment_id)
+            return response, 201
+        return 'You dont have rights', 402
 
     @classmethod
     def create_equipment(cls, data):
-        user = cls._get_user(g.user_id)
-        trip = cls._get_trip(data['trip_id'])
-        user_has_role = data.get('role_id') in (role.id for role in user.roles)
-        is_item_owner = data.get('owner_id') == user.user_id
-        is_admin = user == trip.admin
-        if user_has_role or is_item_owner or is_admin:
+        if cls._user_has_privileges(data['trip_id'], data.get('role_id'), data.get('owner_id')):
             response = current_app.models.Equipment.create_equipment(data)
-        else:
-            response = 'You dont have rights'
-        return response, 201
+            return response, 201
+        return 'You dont have rights', 402
